New version of DEP 1.
This is a more or less complete rewrite of DEP 1. I started with PEP 1,
and made as few changes as possible to modify it to fit our workflow.
The changes from PEP 1 are summarized at the bottom of the new DEP 1.

Thanks to Carl, Jannis, and Andrew for their input on this.

This commit also starts creating the new directory structure for DEPs,
and moves DEP 1 accordingly.