Remove implementation idea for supporting unnamed capture groups.
The proposed syntax maps to named capture groups, and it does not make
sense to also start supporting unnamed capture groups for this.

Also, it is my personal expectation that once this new-style format
becomes popular enough, usage of unnamed capture groups can eventually
be deprecated. (And maybe even should!, but that's talk for the future.)