Added a new technical governance DEP.
This is a simplification and reduction of the technical governance.
It is a merge of DEPs 10 and 12, acting as a single source of truth
for governance.

Add bugfix release mention.

Removes a lot of the extra context around the bugfix being
for a feature release. It's not referenced anywhere else and
may be noise.

Consolidate the steering council voting eligibility revocations.

This is a really long sentence now.

Added additional links for DEP process and new-features repo.

Allowed revisiting vetoed discussions and features at SC discretion.

Allowed DEPs during Steering Council elections

Add specific rationales for the changes to the relevant section.

Remove initial list of definitions in favor of links in the document.

Removed the CLA reference.

Remove the phrase minor consensus.

Add a reporting section that documents what reporting is occurring today.

Replace all references to the DSF secretary with the DSF Board.

Remove the paragraph about revoking the ability to vote.

Remove ability for board members to run for the steering council.

Add updated version of steering council eligibility requirements.

Remove voting as the steering council's decision making process.

Expand the technical direction section to call out ability to use DEPs
for things without using the new features repo.